+++ go-app/frontend/src/tabs/main/MainTabControllerProfileSelecor.tsx
@@ -1,5 +1,6 @@
-import { UseFormReturn } from "react-hook-form";
+import { Controller, UseFormReturn } from "react-hook-form";
 import { main } from "../../../wailsjs/go/models";
+import { useCallback } from "react";
 
 type Props = {
   form: UseFormReturn<{
@@ -21,6 +22,17 @@ type Props = {
   ) => void;
 };
 
+const updatedAtFormatter = new Intl.DateTimeFormat(undefined, {
+  dateStyle: "medium",
+  timeStyle: "medium",
+});
+
+const unfocus = () => {
+  if (document.activeElement && document.activeElement instanceof HTMLElement) {
+    document.activeElement.blur();
+  }
+};
+
 export function MainTabControllerProfileSelector({
   form,
   controller,
@@ -32,72 +44,109 @@ export function MainTabControllerProfileSelector({
   onOpenProfileForController,
   onDeleteProfileForController,
 }: Props) {
-  const { register, watch } = form;
+  const { watch, control } = form;
   const selectedProfile = watch(`profiles.${controller.GUID}`);
+  const supportedProfiles = profiles?.filter(
+    (profile) => !profile.UsbID || profile.UsbID === controller.UsbID,
+  );
+  const unsupportedProfiles = profiles?.filter(
+    (profile) => profile.UsbID && profile.UsbID !== controller.UsbID,
+  );
+
+  const unfocusHandlerFactory = useCallback((func: () => void) => {
+    return () => {
+      func();
+      setTimeout(unfocus, 0);
+    };
+  }, []);
 
   return (
     <div key={controller.GUID} className="flex flex-row gap-2">
-      <fieldset className="fieldset grow">
-        <legend className="fieldset-legend">
-          Select profile for {controller.Name}
-        </legend>
-        <select
-          className="select w-full"
-          {...register(`profiles.${controller.GUID}`)}
-        >
-          <option selected value="">
-            None
-          </option>
-          {profiles
-            ?.filter(
-              (profile) => !profile.UsbID || profile.UsbID === controller.UsbID,
-            )
-            .map((profile) => (
-              <option key={profile.Name} value={profile.Name}>
-                {profile.Name}
-              </option>
-            ))}
-          {profiles
-            ?.filter(
-              (profile) => profile.UsbID && profile.UsbID !== controller.UsbID,
-            )
-            .map((profile) => (
-              <option key={profile.Name} value={profile.Name} disabled>
-                {profile.Name} (Not enabled for controller)
-              </option>
-            ))}
-        </select>
-        {/* <span className="label">
-          Auto-detect only works for certain supported controllers
-        </span> */}
-      </fieldset>
-      <div className="dropdown dropdown-end mt-[34px]">
+      <Controller
+        control={control}
+        name={`profiles.${controller.GUID}`}
+        render={({ field }) => (
+          <div className="grow dropdown dropdown-start">
+            <div tabIndex={0} role="button" className="select w-full">
+              {selectedProfile ?? "Select profile"}
+            </div>
+            <div className="dropdown-content shadow-sm max-h-[50dvh] overflow-auto w-full">
+              <ul className="menu w-full bg-base-300 rounded-box p-2">
+                {supportedProfiles.map((profile) => (
+                  <li key={profile.Name}>
+                    <button
+                      className="grid grid-cols-1 grid-flow-row auto-rows-max gap-0"
+                      onClick={unfocusHandlerFactory(() => {
+                        field.onChange(profile.Name);
+                      })}
+                    >
+                      <span>{profile.Name}</span>
+                      <span className="text-base-content/30 text-xs">
+                        Last updated:{" "}
+                        {updatedAtFormatter.format(
+                          new Date(profile.Metadata.UpdatedAt),
+                        )}
+                      </span>
+                      {!!profile.Metadata.Warnings.length &&
+                        profile.Metadata.Warnings.map((warning) => (
+                          <div
+                            key={warning}
+                            role="alert"
+                            className="alert alert-soft alert-warning my-2 p-2"
+                          >
+                            {warning}
+                          </div>
+                        ))}
+                    </button>
+                  </li>
+                ))}
+                {unsupportedProfiles.map((profile) => (
+                  <li key={profile.Name} className="menu-disabled">
+                    <button className="grid grid-cols-1 grid-flow-row auto-rows-max gap-0">
+                      <span>{profile.Name}</span>
+                      <span className="text-base-content/30 text-xs">
+                        Disabled for controller
+                      </span>
+                    </button>
+                  </li>
+                ))}
+              </ul>
+            </div>
+          </div>
+        )}
+      />
+
+      <div className="dropdown dropdown-end">
         <div tabIndex={0} role="button" className="btn">
           More
         </div>
         <ul
-          tabIndex={0}
+          tabIndex={-1}
           className="dropdown-content menu bg-base-100 rounded-box z-1 w-52 p-2 shadow-sm"
         >
           <li>
-            <button onClick={onReloadConfiguration}>
+            <button onClick={unfocusHandlerFactory(onReloadConfiguration)}>
               Reload configuration
             </button>
           </li>
           <li>
-            <button onClick={onBrowseConfiguration}>
+            <button onClick={unfocusHandlerFactory(onBrowseConfiguration)}>
               Browse configuration
             </button>
           </li>
           <li>
-            <button onClick={() => onCreateProfile(controller)}>
+            <button
+              onClick={unfocusHandlerFactory(() => onCreateProfile(controller))}
+            >
               Create new profile
             </button>
           </li>
           <li>
             <button
               disabled={!selectedProfile}
-              onClick={() => onSaveControllerProfileForSharing(controller)}
+              onClick={unfocusHandlerFactory(() =>
+                onSaveControllerProfileForSharing(controller),
+              )}
               className="disabled:opacity-50 disabled:pointer-events-none"
             >
               Save profile for sharing
@@ -106,7 +155,9 @@ export function MainTabControllerProfileSelector({
           <li>
             <button
               disabled={!selectedProfile}
-              onClick={() => onOpenProfileForController(controller)}
+              onClick={unfocusHandlerFactory(() =>
+                onOpenProfileForController(controller),
+              )}
               className="disabled:opacity-50 disabled:pointer-events-none"
             >
               Open profile in builder
@@ -115,7 +166,9 @@ export function MainTabControllerProfileSelector({
           <li>
             <button
               disabled={!selectedProfile}
-              onClick={() => onDeleteProfileForController(controller)}
+              onClick={unfocusHandlerFactory(() =>
+                onDeleteProfileForController(controller),
+              )}
               className="disabled:opacity-50 disabled:pointer-events-none"
             >
               Delete profile