+++ go-app/frontend/wailsjs/go/main/App.d.ts
@@ -6,6 +6,8 @@ export function ClearProfile(arg1:string):Promise<void>;
 
 export function DeleteProfile(arg1:string):Promise<void>;
 
+export function GetAlwaysOnTop():Promise<boolean>;
+
 export function GetCabControlState():Promise<main.Interop_Cab_ControlState>;
 
 export function GetControllerConfiguration(arg1:string):Promise<main.Interop_ControllerConfiguration>;
@@ -55,6 +57,8 @@ export function SaveLogs(arg1:Array<string>):Promise<void>;
 export function SaveProfileForSharing(arg1:string,arg2:string):Promise<void>;
 
 export function SelectProfile(arg1:string,arg2:string):Promise<void>;
+
+export function SetAlwaysOnTop(arg1:boolean):Promise<void>;
 
 export function SetLastInstalledModVersion(arg1:string):Promise<void>;